Исправления:
- Убрал лишние свойства - userDao и gson.
- Поле dbServiceClient в ClientsWebServerHomework теперь инициализируется через конструктор.
- Инициализацию БД перенёс в класс Homework.
- Убрал инициализацию начальными данными.
- Убрал сервлет ClientsApiServlet, теперь добавление клиента через метод POST в /clients.
- Добавил закрывающие теги для TR/TD в шаблон clients.html.
- Добавил вывод сообщения на странице с клиентами в случае, если переданы пустые значения для имени или адреса.
