+++ hw12-webserver/webserver/src/main/java/ru/otus/Homework.java
@@ -1,7 +1,14 @@
 package ru.otus;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
+import org.hibernate.cfg.Configuration;
+import ru.otus.core.repository.DataTemplateHibernate;
+import ru.otus.core.repository.HibernateUtils;
+import ru.otus.core.sessionmanager.TransactionManagerHibernate;
+import ru.otus.crm.model.AddressDataSet;
+import ru.otus.crm.model.Client;
+import ru.otus.crm.model.PhoneDataSet;
+import ru.otus.crm.service.DBServiceClient;
+import ru.otus.crm.service.DbServiceClientImpl;
 import ru.otus.dao.InMemoryUserDao;
 import ru.otus.dao.UserDao;
 import ru.otus.server.ClientsWebServer;
@@ -20,30 +27,36 @@ import ru.otus.services.UserAuthServiceImpl;
     // список клиентов
     http://localhost:8080/clients
 
-    // АПИ
-    http://localhost:8080/api/clients
-        реализовано только add
-
 */
 
 public class Homework {
     private static final int WEB_SERVER_PORT = 8080;
     private static final String TEMPLATES_DIR = "/templates/";
+    public static final String HIBERNATE_CFG_FILE = "hibernate.cfg.xml";
 
     public static void main(String[] args) throws Exception {
         UserDao userDao = new InMemoryUserDao();
-        Gson gson = new GsonBuilder()
-                .serializeNulls()
-                .setPrettyPrinting()
-                .excludeFieldsWithoutExposeAnnotation()
-                .create();
         TemplateProcessor templateProcessor = new TemplateProcessorImpl(TEMPLATES_DIR);
         UserAuthService authService = new UserAuthServiceImpl(userDao);
+        DBServiceClient dbServiceClient = getDBServiceClient();
 
         ClientsWebServer usersWebServer = new ClientsWebServerHomework(WEB_SERVER_PORT,
-                userDao, gson, templateProcessor, authService);
+                templateProcessor, authService, dbServiceClient);
 
         usersWebServer.start();
         usersWebServer.join();
+    }
+
+    private static DbServiceClientImpl getDBServiceClient() {
+        var configuration = new Configuration().configure(HIBERNATE_CFG_FILE);
+
+        var sessionFactory = HibernateUtils.buildSessionFactory(configuration,
+                Client.class,
+                AddressDataSet.class,
+                PhoneDataSet.class);
+
+        var transactionManager = new TransactionManagerHibernate(sessionFactory);
+        var clientTemplate = new DataTemplateHibernate<>(Client.class);
+        return new DbServiceClientImpl(transactionManager, clientTemplate);
     }
 }
