+++ hw12-webserver/webserver/src/main/java/ru/otus/server/ClientsWebServerHomework.java
@@ -1,7 +1,5 @@
 package ru.otus.server;
 
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.handler.HandlerList;
@@ -9,52 +7,38 @@ import org.eclipse.jetty.server.handler.ResourceHandler;
 import org.eclipse.jetty.servlet.FilterHolder;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
-import org.hibernate.cfg.Configuration;
-import ru.otus.core.repository.DataTemplateHibernate;
-import ru.otus.core.repository.HibernateUtils;
-import ru.otus.core.sessionmanager.TransactionManagerHibernate;
-import ru.otus.crm.dbmigrations.MigrationsExecutorFlyway;
-import ru.otus.crm.model.AddressDataSet;
-import ru.otus.crm.model.Client;
-import ru.otus.crm.model.PhoneDataSet;
 import ru.otus.crm.service.DBServiceClient;
-import ru.otus.crm.service.DbServiceClientImpl;
-import ru.otus.dao.UserDao;
 import ru.otus.helpers.FileSystemHelper;
 import ru.otus.services.TemplateProcessor;
 import ru.otus.services.UserAuthService;
-import ru.otus.servlet.*;
+import ru.otus.servlet.AuthorizationFilter;
+import ru.otus.servlet.ClientsServlet;
+import ru.otus.servlet.LoginServlet;
 
-import java.lang.reflect.Type;
-import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.stream.Collectors;
 
 public class ClientsWebServerHomework implements ClientsWebServer {
     private static final String START_PAGE_NAME = "index.html";
     private static final String COMMON_RESOURCES_DIR = "static";
-    public static final String HIBERNATE_CFG_FILE = "hibernate.cfg.xml";
 
-    private final UserDao userDao;
-    private final Gson gson;
     protected final TemplateProcessor templateProcessor;
     private final Server server;
     private final UserAuthService authService;
-    private DbServiceClientImpl dbServiceClient;
+    private final DBServiceClient dbServiceClient;
 
-    public ClientsWebServerHomework(int port, UserDao userDao, Gson gson, TemplateProcessor templateProcessor, UserAuthService authService) {
-        this.userDao = userDao;
-        this.gson = gson;
+    public ClientsWebServerHomework(int port,
+                                    TemplateProcessor templateProcessor,
+                                    UserAuthService authService,
+                                    DBServiceClient dbServiceClient) {
         this.templateProcessor = templateProcessor;
         server = new Server(port);
         this.authService = authService;
+        this.dbServiceClient = dbServiceClient;
     }
 
     @Override
     public void start() throws Exception {
         if (server.getHandlers().length == 0) {
-            InitDB();
-            //initData(); // для теста
             initContext();
         }
         server.start();
@@ -70,10 +54,6 @@ public class ClientsWebServerHomework implements ClientsWebServer {
         server.stop();
     }
 
-    public DBServiceClient getDbServiceClient() {
-        return this.dbServiceClient;
-    }
-
     protected Handler applySecurity(ServletContextHandler servletContextHandler, String... paths) {
         servletContextHandler.addServlet(new ServletHolder(new LoginServlet(templateProcessor, authService)), "/login");
         AuthorizationFilter authorizationFilter = new AuthorizationFilter();
@@ -81,17 +61,15 @@ public class ClientsWebServerHomework implements ClientsWebServer {
         return servletContextHandler;
     }
 
-    private Server initContext() {
-
+    private void initContext() {
         ResourceHandler resourceHandler = createResourceHandler();
         ServletContextHandler servletContextHandler = createServletContextHandler();
 
         HandlerList handlers = new HandlerList();
         handlers.addHandler(resourceHandler);
-        handlers.addHandler(applySecurity(servletContextHandler, "/clients", "/api/clients/*"));
+        handlers.addHandler(applySecurity(servletContextHandler, "/clients"));
 
         server.setHandler(handlers);
-        return server;
     }
 
     private ResourceHandler createResourceHandler() {
@@ -104,46 +82,7 @@ public class ClientsWebServerHomework implements ClientsWebServer {
 
     private ServletContextHandler createServletContextHandler() {
         ServletContextHandler servletContextHandler = new ServletContextHandler(ServletContextHandler.SESSIONS);
-        servletContextHandler.addServlet(new ServletHolder(new ClientsServlet(templateProcessor, dbServiceClient, gson)), "/clients");
-        servletContextHandler.addServlet(new ServletHolder(new ClientsApiServlet(dbServiceClient, gson)), "/api/clients/*");
+        servletContextHandler.addServlet(new ServletHolder(new ClientsServlet(templateProcessor, dbServiceClient)), "/clients");
         return servletContextHandler;
-    }
-
-    private void InitDB() {
-        var configuration = new Configuration().configure(HIBERNATE_CFG_FILE);
-
-        var dbUrl = configuration.getProperty("hibernate.connection.url");
-        var dbUserName = configuration.getProperty("hibernate.connection.username");
-        var dbPassword = configuration.getProperty("hibernate.connection.password");
-
-        var sessionFactory = HibernateUtils.buildSessionFactory(configuration,
-                Client.class,
-                AddressDataSet.class,
-                PhoneDataSet.class);
-
-        var transactionManager = new TransactionManagerHibernate(sessionFactory);
-        var clientTemplate = new DataTemplateHibernate<>(Client.class);
-        this.dbServiceClient = new DbServiceClientImpl(transactionManager, clientTemplate);
-    }
-
-    private void initData() {
-        for (var i = 1; i <= 2; i++) {
-            var clientAddress = new AddressDataSet();
-            var client = new Client("Client #" + i);
-            clientAddress.setStreet("#" + i  + " Client address", client);
-            var clientPhones = new ArrayList<PhoneDataSet>();
-            for (int j = 1; j <= 5; j++) {
-                clientPhones.add(new PhoneDataSet("number" + j * i, client));
-            }
-
-            client.setAddress(clientAddress);
-            client.setPhones(clientPhones);
-
-            dbServiceClient.saveClient(client);
-        }
-
-        var clients = dbServiceClient.findAll();
-        Type dataType = new TypeToken<ArrayList<Client>>(){}.getType();
-        System.out.println(gson.toJson(clients, dataType));
     }
 }
