+++ hw12-webserver/webserver/src/main/resources/templates/clients.html
@@ -1,6 +1,10 @@
 <html xmlns="http://www.w3.org/1999/xhtml">
 <head>
     <script>
+        function hideAlert() {
+            document.getElementById("alert").style.display = "none";
+        }
+
         function addPhone() {
             const phones = document.getElementById("phones");
             const phone = document.getElementById("phone");
@@ -40,18 +44,19 @@
 </head>
 
 
-<form action="api/clients/add" method="post">
+<form action="clients" method="post">
+    <input type="hidden" name ="action" value="add">
     <table>
         <tr>
-            <td>Клиент:
-            <td><input id="name" name="name" type="text">
+            <td>Клиент:</td>
+            <td><input id="name" name="name" type="text"></td>
         </tr>
         <tr>
-            <td>Адрес:
-            <td><input id="address" name="address" type="text">
+            <td>Адрес:</td>
+            <td><input id="address" name="address" type="text"></td>
         </tr>
         <tr>
-            <td>Телефоны:
+            <td>Телефоны:</td>
             <td>
                 <select id="phones" name="phones" size = 5 style="width:200px" multiple>
                 </select><br>
@@ -59,23 +64,31 @@
                 <input type="button" value="добавить" onclick="addPhone()">&nbsp;
                 <input type="button" value="удалить" onclick="deletePhone()">&nbsp;
                 <input type="button" value="очистить" onclick="deleteAllPhones()">
+            </td>
     </table>
     <input type="submit" value="Добавить клиента" onclick="selectAll()">
 </form>
 
+<#if alertMessage??>
+<div id="alert" style="color: red">${alertMessage}<input type="button" onclick="hideAlert()" value="Ок"></div>
+</#if>
+
 <h4>Клиенты</h4>
 <table border="1" style="border-width:1; border-style:solid; border-collapse:collapse">
     <tr>
-        <td>Client id<td>Client name<td>Client address<td>Client phones
-        <#list allClients?eval as client>
+        <td>Client id</td><td>Client name</td><td>Client address</td><td>Client phones</td>
+    </tr>
+    <#list allClients as client>
     <tr>
-        <td>${client.id}
-        <td>${client.name}
-        <td>${client.address.street}
+        <td>${client.id}</td>
+        <td>${client.name}</td>
+        <td>${client.address.street}</td>
         <td>
             <#list client.phones as phone>
             <div>${phone.number}</div>
         </#list>
+        </td>
+    </tr>
     </#list>
 </table>
 </body>
